Update previews in composer
